add middleware author of book, add middleware in access list(for library), add access by link

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AccessController;
 use App\Http\Controllers\BookController;
 use App\Http\Controllers\CommentController;
 use App\Http\Controllers\HomeController;
@@ -21,20 +22,27 @@ Route::get('/', function () {
     return redirect(route('profile.index'));
 });
 
-Route::prefix('comments')->controller(CommentController::class)->group(function () {
-    Route::get('/', 'index')->name('comments.index');
-    Route::post('/', 'store')->name('comments.store');
-    Route::get('/{comment}/answer', 'answer')->name('comments.answer');
-    Route::delete('/{comment}', 'delete')->name('comments.delete');
+Route::prefix('access')->controller(AccessController::class)->middleware('auth')->group(function () {
+    Route::post('/add', 'add')->name('access.list.add');
+    Route::post('/delete', 'delete')->name('access.list.delete');
 });
 
-Route::prefix('books')->controller(BookController::class)->group(function () {
-    Route::get('/create', 'create')->name('books.create')->middleware('auth');
+Route::prefix('comments')->controller(CommentController::class)->group(function () {
+    Route::get('/', 'index')->name('comments.index');
+    Route::post('/', 'store')->name('comments.store')->middleware('auth');
+    Route::get('/{comment}/answer', 'answer')->name('comments.answer')->middleware('auth');
+    Route::delete('/{comment}', 'delete')->name('comments.delete')->middleware('auth');
+});
+
+Route::prefix('books')->controller(BookController::class)->middleware('auth')->group(function () {
+    Route::get('/create', 'create')->name('books.create');
+    Route::get('/link/{link}', 'showByLink')->name('books.showByLink');
     Route::post('/', 'store')->name('books.store');
-    Route::get('/{book}', 'show')->name('books.show');
-    Route::get('/{book}/edit', 'edit')->name('books.edit');
-    Route::patch('/{book}', 'update')->name('books.update');
-    Route::delete('/{book}', 'delete')->name('books.delete');
+    Route::get('/{book}', 'show')->name('books.show')->middleware('access.list');
+    Route::get('/{book}/edit', 'edit')->name('books.edit')->middleware('book.author');
+    Route::patch('/{book}', 'update')->name('books.update')->middleware('book.author');
+    Route::delete('/{book}', 'delete')->name('books.delete')->middleware('book.author');
+    Route::post('/{book}/make_link','makeLink')->name('books.make.link')->middleware('book.author');
 });
 
 Route::prefix('/profiles')->controller(ProfileController::class)->group(function () {
